# openapi.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    Item:
      properties:
        description:
          anyOf:
          - type: string
          title: Description
        item_id:
          title: Item Id
          type: string
        name:
          title: Name
          type: string
      required:
      - item_id
      - name
      title: Item
      type: object
    UserInfo:
      properties:
        roles:
          items:
            type: string
          title: Roles
          type: array
        user_id:
          title: User Id
          type: string
        username:
          title: Username
          type: string
      required:
      - user_id
      - username
      - roles
      title: UserInfo
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    APIKeyHeader:
      in: header
      name: X-API-Key
      type: apiKey
info:
  description: "\u0424\u0456\u043A\u0442\u0438\u0432\u043D\u0438\u0439 API \u0437\
    \ \u0430\u0432\u0442\u043E\u0440\u0438\u0437\u0430\u0446\u0456\u0454\u044E X-API-Key.\
    \ \u041F\u0435\u0440\u0435\u0432\u0456\u0440\u0442\u0435 /openapi.json \u0442\u0430\
    \ /docs."
  title: Dummy Product API
  version: 1.0.0
openapi: 3.1.0
paths:
  /:
    get:
      operationId: root__get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Root
  /items/{item_id}:
    get:
      operationId: get_item_info_items__item_id__get
      parameters:
      - in: path
        name: item_id
        required: true
        schema:
          title: Item Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyHeader: []
      summary: Get Item Info
      tags:
      - Items
  /users/{user_id}:
    get:
      operationId: get_user_info_users__user_id__get
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyHeader: []
      summary: Get User Info
      tags:
      - Users